HeroKuWebSite/requirements.txt
Flask==1.1.4
pandas==1.1.5
numpy==1.19.5
gunicorn==20.1.0
line-bot-sdk==1.19.0
beautifulsoup4==4.6.3
requests==2.23.0
regex==2019.12.20
bert-for-tf2==0.14.9
sentencepiece==0.1.69
transformers==4.10.0
tensorflow==2.6.0
tensorflow-hub==0.12.0